Issue #16: Split adapters to separate modules.
