+++ tessumod/src/scripts/client/gui/mods/tessumod/adapters/__init__.py
@@ -1,0 +1,11 @@
+from . import settings
+from . import teamspeak
+from . import usercache
+from . import wotgame
+
+__all__ = [
+	"settings",
+	"teamspeak",
+	"usercache",
+	"wotgame"
+]
